Ignore blur on menu enter/leave instead of the individual items #211 #222
IE11 (and probably older versions) trigger a click event when you grab the
scrollbar thumb. Since the scrollbar isn't part of any individual item within
the menu, we promptly closed it, thinking the click happened outside
the dropdown. This commit changes the logic to instead capture focus while the
user is hovering over the menu, keeping the input focused until the user either
clicks outside or selects an item.