Delay re-focus until all actions have been processed (#240)
This is a fix for when trying to select an item while the input was off-screen:
The focus was being reset immediately after it had been lost, causing the view-
port to scroll to the input field before the item's onClick handler had been
processed. This was a problem because if the cursor moved as much as one pixel
in the meantime it would change the highlightedIndex, causing a re-render
(because it would now likely be hovering over a different item) and aborting
the chain of actions (i.e. the item would not be selected).