Simplify autoHighlight logic
The auto highlight logic now happens when the component receives new props,
which is the only  - time apart from using the arrow keys or the mouse - that
the highlighted index changes. The previous logic was from when props.value was
unmanaged, and needed to hook into several places to stay synced.